2012.10.01 Windows 2012 Server 64 bit Jelentéktelen warningok javítása. MSC fordítási ociók: -G5 -GX megszűnt. MSC linkelési opciók:  bufferoverflowu.lib javítva: gtk_text_iter_apix.cpp, gtk_tree_iter_apix.cpp

+++ ccclib/src/stat.cpp
@@ -77,9 +77,9 @@ static void createStatEntry(struct stat *buf)
         number(0); //Windowson nincs
     #endif
  
-    number(buf->st_atime);
-    number(buf->st_mtime);
-    number(buf->st_ctime);
+    number((double)buf->st_atime);
+    number((double)buf->st_mtime);
+    number((double)buf->st_ctime);
 
     array(N_STAT);
 }
